25
25

+++ SWiOS/ActivityProduct.h
@@ -13,7 +13,7 @@
 @property(nonatomic,copy)NSString *productCode;
 @property(nonatomic,copy)NSString *productName;
 @property(nonatomic,strong)NSNumber *rushPrice;
-@property(nonatomic,assign)NSInteger rushQuantity;
+@property(nonatomic,assign)int rushQuantity;
 @property(nonatomic,copy)NSString *picUrl1;
 @property (assign,nonatomic) NSInteger buyCount;
 @end
